backend/pyproject.toml
[project]
name = "stock-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.124.4",
    "langchain[openai]>=0.0.27",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.33.0",
    "yfinance>=0.2.66",
]
